requirements.txt
django==5.1.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
asgiref==3.8.1
sqlparse==0.5.1
gunicorn==23.0.0
packaging==24.1 
flake8==7.1.1
mccabe==0.7.0
pycodestyle==2.12.1
pyflakes==3.2.0
iniconfig==2.0.0 
pluggy==1.5.0 
pytest==8.3.3
